Split out stream writer tests into separate files
